Adicionando uma nova tabela de usuarios e adicionando um controlador de erros

+++ src/main/java/com/apilivros/livros/controller/ApiDeLivrosController.java
@@ -1,8 +1,9 @@
 package com.apilivros.livros.controller;
 
-import com.apilivros.livros.model.Livros;
+import com.apilivros.livros.exceptions.UserNotFound;
+import com.apilivros.livros.tabelas.Livros;
 import com.apilivros.livros.repositoty.ApiDeLivrosRepository;
-import org.springframework.http.ResponseEntity;
+import org.springframework.http.HttpStatus;
 import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 
@@ -20,7 +21,9 @@ public class ApiDeLivrosController {
         this.repository = repository;
     }
 
+
     //Colocando um array de objetos
+    @CrossOrigin(origins = "http://localhost:5173")
     @PostMapping()
     public List<Livros> addLivros(@RequestBody List<Livros> livros) {
 
@@ -33,27 +36,47 @@ public class ApiDeLivrosController {
         return livros;
     }
 
+    @CrossOrigin(origins = "http://localhost:5173")
     @GetMapping("/{id}")
     public Optional<Livros> get (@PathVariable("id") String id) {
+
         Optional<Livros> livro = repository.findById(id);
+        if (livro.isEmpty()) {
+            throw new UserNotFound();
+        }
+
         return livro;
     }
 
+    @CrossOrigin(origins = "http://localhost:5173")
     @GetMapping()
     public List<Livros> getByNome (@RequestParam(value = "nome", required = false) String nome) {
+        List<Livros> livros = repository.findByNomeContaining(nome);
+        if (livros.isEmpty()) {
+            throw new UserNotFound();
+        }
         if (!StringUtils.hasText(nome)) {
             return repository.findAll();
         }
-        return  repository.findByNomeContaining(nome);
+
+        return  livros;
     }
 
+    @CrossOrigin(origins = "http://localhost:5173")
     @DeleteMapping("/{id}")
     public void delet (@PathVariable("id") String id) {
+        if (repository.findById(id).isEmpty()) {
+            throw new UserNotFound();
+        }
         repository.deleteById(id);
     }
 
+    @CrossOrigin(origins = "http://localhost:5173")
     @PutMapping("/{id}")
     public void alterar (@RequestBody Livros livro, @PathVariable("id") String id) {
+        if (repository.findById(id).isEmpty()) {
+            throw new UserNotFound();
+        }
         livro.setIdLivro(id);
         repository.save(livro);
     }
